// apps/vendor/src/services/models/NotificationView.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { User } from './User';

export type NotificationView = {
    id?: string;
    title?: string | null;
    message?: string | null;
    url?: string | null;
    userId?: string;
    user?: User;
    isRead?: boolean;
    dateCreated?: string;
};