// apps/vendor/src/services/models/NotificationViewPagedCollection.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { Link } from './Link';
import type { NotificationView } from './NotificationView';

export type NotificationViewPagedCollection = {
    offset?: number | null;
    limit?: number | null;
    size?: number;
    first?: Link;
    previous?: Link;
    next?: Link;
    last?: Link;
    self?: Link;
    value?: Array<NotificationView> | null;
};